worker/requirements.txt
joblib
redis
python-dotenv

tensorflow
matplotlib
keras-cv
opencv-python


kafka-python
